+++ RELEASE_NOTES.md
@@ -1,6 +1,11 @@
 Release Notes
 =============
 
+## vNext
+* Galleries: Create Galleries for sharing VM images.
+* Gallery Images: Create images in a gallery that can be used to create VMs.
+* Image Templates: Templates for building and publishing VM images.
+
 ## 1.7.17
 * Container Apps: Fix scaler spelling
 * Disks: Create managed disks